+++ scripts/winesapos-setup.sh
@@ -668,18 +668,6 @@ if [ $? -eq 0 ]; then
     sudo systemctl enable --now zerotier-one
 fi
 
-kdialog --title "winesapOS First-Time Setup" --yesno "Do you want to install support for the Bcachefs file system?"
-if [ $? -eq 0 ]; then
-    kdialog_dbus=$(kdialog --title "winesapOS First-Time Setup" --progressbar "Please wait for Bcachefs support to be installed..." 4 | cut -d" " -f1)
-    qdbus ${kdialog_dbus} /ProgressDialog Set org.kde.kdialog.ProgressDialog value 1
-    ${CMD_YAY_INSTALL[*]} linux-bcachefs-git
-    qdbus ${kdialog_dbus} /ProgressDialog Set org.kde.kdialog.ProgressDialog value 2
-    ${CMD_YAY_INSTALL[*]} linux-bcachefs-git-headers
-    qdbus ${kdialog_dbus} /ProgressDialog Set org.kde.kdialog.ProgressDialog value 3
-    ${CMD_YAY_INSTALL[*]} bcachefs-tools-git
-    qdbus ${kdialog_dbus} /ProgressDialog org.kde.kdialog.ProgressDialog.close
-fi
-
 if [[ "${WINESAPOS_IMAGE_TYPE}" != "secure" ]]; then
     kdialog --title "winesapOS First-Time Setup" --yesno "Do you want to change your password?"
     if [ $? -eq 0 ]; then