+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Change Pacman to not ignore any package updates by default.
 - Change scripts to use Bash instead of Zsh.
 - Change 'linux-lts' to be Linux 6.6 instead of 6.1.
+- Change Bcachefs to be installed using the official packages instead of from the AUR.
 
 ## 3.4.0 - The Game Mode Update